requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
openai>=1.12.0
chromadb>=0.4.22
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
streamlit>=1.31.0
httpx>=0.26.0
tiktoken>=0.5.2
pytest>=7.0.0
pytest-asyncio>=0.23.0
redis>=5.0.0
